Move looping tests out of race detector
